opencv-service/requirements.txt
flask
flask-cors
opencv-python-headless
numpy
requests
ultralytics
scikit-learn
pandas
mysql-connector-python
python-dotenv
